requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp_retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
av==12.3.0
blinker==1.9.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.3.0
clicksend-client==5.0.78
colorzero==2.0
distro==1.8.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
gpiozero==2.0.1
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.3
jsonschema==4.10.3
lgpio==0.2.2.0
libarchive-c==2.9
MarkupSafe==3.0.2
multidict==6.6.4
numpy==1.24.2
olefile==0.46
paho-mqtt==2.1.0
picamera2==0.3.31
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
propcache==0.3.2
pycryptodomex==3.11.0
PyJWT==2.10.1
PyOpenGL==3.1.6
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-prctl==1.8.1
PyYAML==6.0.2
requests==2.28.1
rfc3987==1.3.8
rpi-lgpio==0.6
simplejpeg==1.8.1
six==1.16.0
smbus2==0.4.2
spidev==3.5
SQLAlchemy==2.0.43
ssh-import-id==5.10
toml==0.10.2
tqdm==4.64.1
twilio==9.8.1
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==1.26.12
videodev2==0.0.4
webcolors==1.11.1
Werkzeug==3.1.3
yarl==1.20.1